feature: implement simple pawn promotion feature
Supports promotion of pawn to queen. Technically  pawn can be promoted to queen, bishop, knight or rook, but Queen is most common in normal play. Additional UI paradigms will be required to allow a user to select a piece type for promotion, and "down promotion" is really only useful to avoid stalemate in some rare situations.